Store BTC keys in PGP-encrypted files; put addresses in the DB.
This moves the code to storing the BTC private keys in per-user
PGP-encrypted files on disk, and stores the addresses (public
data) unencrypted in the DB. It pulls in openpgp to do the
encryption. This change ensures that, aside from in memory (and
swap in worst case scenario), the BTC private keys never exist in
an unencrypted state on the system. Furthermore, because of how
public key encryption works, the PGP private keys for decrypting
the files storing the BTC private keys do not need to be present
on the server running the code, adding another layer of security.